Analytics + Permission + Cookies Bug Fixes (#110)
* Removed setting/accessing cookies in client

* Fixed volunteer info not loading + image link updat fail if empty

* Fixed analytics retrieval

+++ src/app/api/volunteer/profile-image/image-link/route.ts
@@ -38,21 +38,23 @@ export const POST = APIWrapper({
     }
 
     const blobName = currentUser!.imageLink.split("/").slice(4).join("/");
-    const sharedKeyCredential = new StorageSharedKeyCredential(
-      accountName,
-      accountKey,
-    );
-    const blobServiceClient = new BlobServiceClient(
-      `https://${accountName}.blob.core.windows.net`,
-      sharedKeyCredential,
-    );
-    const containerClient = blobServiceClient.getContainerClient(containerName);
-    const blobClient = containerClient.getBlockBlobClient(blobName);
-    const options: BlobDeleteOptions = {
-      deleteSnapshots: "include" as DeleteSnapshotsOptionType,
-    };
-    await blobClient.deleteIfExists(options);
-
+    if (blobName !== "") {
+      const sharedKeyCredential = new StorageSharedKeyCredential(
+        accountName,
+        accountKey,
+      );
+      const blobServiceClient = new BlobServiceClient(
+        `https://${accountName}.blob.core.windows.net`,
+        sharedKeyCredential,
+      );
+      const containerClient =
+        blobServiceClient.getContainerClient(containerName);
+      const blobClient = containerClient.getBlockBlobClient(blobName);
+      const options: BlobDeleteOptions = {
+        deleteSnapshots: "include" as DeleteSnapshotsOptionType,
+      };
+      await blobClient.deleteIfExists(options);
+    }
     const user = await updateVolunteer(email, { imageLink: newImageLink });
     return user;
   },
